fix: Moved FirebaseError interface to app module
