+++ src/firebase-namespace-api.ts
@@ -29,85 +29,6 @@ import { App as AppCore, AppOptions } from './app/index';
 
 export * from './app/index';
 
-/**
- * `FirebaseError` is a subclass of the standard JavaScript `Error` object. In
- * addition to a message string and stack trace, it contains a string code.
- */
-export interface FirebaseError {
-
-  /**
-   * Error codes are strings using the following format: `"service/string-code"`.
-   * Some examples include `"auth/invalid-uid"` and
-   * `"messaging/invalid-recipient"`.
-   *
-   * While the message for a given error can change, the code will remain the same
-   * between backward-compatible versions of the Firebase SDK.
-   */
-  code: string;
-
-  /**
-   * An explanatory message for the error that just occurred.
-   *
-   * This message is designed to be helpful to you, the developer. Because
-   * it generally does not convey meaningful information to end users,
-   * this message should not be displayed in your application.
-   */
-  message: string;
-
-  /**
-   * A string value containing the execution backtrace when the error originally
-   * occurred.
-   *
-   * This information can be useful to you and can be sent to
-   * {@link https://firebase.google.com/support/ Firebase Support} to help
-   * explain the cause of an error.
-   */
-  stack?: string;
-
-  /**
-   * @return A JSON-serializable representation of this object.
-   */
-  toJSON(): object;
-}
-
-/**
- * Composite type which includes both a `FirebaseError` object and an index
- * which can be used to get the errored item.
- *
- * @example
- * ```javascript
- * var registrationTokens = [token1, token2, token3];
- * admin.messaging().subscribeToTopic(registrationTokens, 'topic-name')
- *   .then(function(response) {
- *     if (response.failureCount > 0) {
- *       console.log("Following devices unsucessfully subscribed to topic:");
- *       response.errors.forEach(function(error) {
- *         var invalidToken = registrationTokens[error.index];
- *         console.log(invalidToken, error.error);
- *       });
- *     } else {
- *       console.log("All devices successfully subscribed to topic:", response);
- *     }
- *   })
- *   .catch(function(error) {
- *     console.log("Error subscribing to topic:", error);
- *   });
- *```
- */
-export interface FirebaseArrayIndexError {
-
-  /**
-   * The index of the errored item within the original array passed as part of the
-   * called API method.
-   */
-  index: number;
-
-  /**
-   * The error object.
-   */
-  error: FirebaseError;
-}
-
 // eslint-disable-next-line @typescript-eslint/no-namespace
 export namespace app {
   /**
