+++ test/unit/storage/storage.spec.ts
@@ -22,7 +22,7 @@ import { expect } from 'chai';
 
 import * as mocks from '../../resources/mocks';
 import { FirebaseApp } from '../../../src/app/firebase-app';
-import { Storage } from '../../../src/storage/storage';
+import { Storage } from '../../../src/storage/index';
 
 describe('Storage', () => {
   let mockApp: FirebaseApp;
